Traceroot/requirements.txt
flask==3.1.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
